ecommerce/apps/accounts/serializers/login.py
from rest_framework_simplejwt.serializers import TokenObtainPairSerializer


class LoginSerializer(TokenObtainPairSerializer):
    def validate(self, attrs):
        data = super().validate(attrs)
        data["user"] = {
            "email": self.user.email,
            "role": self.user.role,
            "id": self.user.id,
        }
        return data
