ecommerce/requirements.txt
django>=5.0
djangorestframework
djangorestframework-simplejwt
django-filter
django-environ
pillow
drf-spectacular # For API documentation
django-cors-headers
